Divided main code from jupyter notebooks Modularized LAB_UW_functions and moved to src/
