+++ src/synthetic_data.py
@@ -1,0 +1,38 @@
+# src/synthetic_data.py
+import numpy as np
+def synthetic_wavelets_in_noise(nsamples: int = 1000, noise_amplitude: float = 0.5, winlen: int = 100,
+                                 square_start: int = 400, amp_square: float = 5, frequency: int = 4,
+                                 amplitude: int = 1) -> np.ndarray:
+    """
+    Generate synthetic wavelets in noise.
+
+    Args:
+        nsamples (int): Number of samples.
+        noise_amplitude (float): Amplitude of the noise.
+        winlen (int): Length of the window.
+        square_start (int): Start index of the square wave.
+        amp_square (float): Amplitude of the square wave.
+        frequency (int): Frequency of the sine wave.
+        amplitude (int): Amplitude of the sine wave.
+
+    Returns:
+        np.ndarray: Synthetic waveform with wavelets in noise.
+    """
+    time = np.linspace(0, 10, nsamples)
+
+    noise = np.random.normal(0, noise_amplitude, nsamples)
+
+    signal = np.zeros(noise.shape)
+
+    mask = amp_square * np.hanning(winlen)
+    signal[square_start:square_start + winlen] = mask
+
+    sin_signal = np.sin(2 * np.pi * frequency * time[square_start:square_start + winlen])
+    signal[square_start:square_start + winlen] = mask * sin_signal
+
+    waveform = signal + noise
+
+    return waveform
+
+# Example usage:
+# waveform = synthetic_wavelets_in_noise()
